rewrite download and perimeters
