+++ packages/ui/src/components/_common/downloads.tsx
@@ -1,11 +1,12 @@
-import { defineComponent, FunctionalComponent, ref } from 'vue'
+import { defineComponent, inject } from 'vue'
 import { useRoute } from 'vue-router'
 import { PureDownloads, Props } from './pure-downloads'
 
-export const Downloads = defineComponent<Omit<Props, 'route'>>({
-  props: ['formats', 'section', 'params'] as unknown as undefined,
+export const Downloads = defineComponent<Omit<Props, 'route' | 'matomo'>>({
+  props: ['formats', 'section'] as unknown as undefined,
   setup(props) {
     const route = useRoute()
-    return () => <PureDownloads {...props} route={route} />
+    const matomo = inject('matomo', undefined)
+    return () => <PureDownloads {...props} route={route} matomo={matomo} />
   }
 })
