+++ packages/ui/src/components/_common/pure-downloads.stories.tsx
@@ -1,5 +1,7 @@
+import { action } from '@storybook/addon-actions'
 import { Meta, Story } from '@storybook/vue3'
 import { RouteLocationNormalized } from 'vue-router'
+import { Store } from 'vuex'
 import { PureDownloads, Props } from './pure-downloads'
 
 const meta: Meta = {
@@ -8,6 +10,7 @@ const meta: Meta = {
 }
 export default meta
 
+const dispatch = action('dispatch')
 export const Default: Story = () => (
   <PureDownloads
     formats={['geojson', 'xlsx']}
