+++ packages/ui/src/components/_common/pure-downloads.tsx
@@ -1,16 +1,17 @@
-import { defineComponent, FunctionalComponent, ref } from 'vue'
-import { RouteLocationNormalized, useRoute } from 'vue-router'
+import { defineComponent, ref } from 'vue'
+import { RouteLocationNormalized } from 'vue-router'
 import { Dropdown } from '../_ui/dropdown'
-import Download from './download.vue'
+import { Download } from './download'
 
 export interface Props {
   formats: string[]
   section: string
   params: any
   route: RouteLocationNormalized
+  matomo?: { trackLink: (url: string, params: string) => void }
 }
 export const PureDownloads = defineComponent<Props>({
-  props: ['route', 'formats', 'section', 'params'] as unknown as undefined,
+  props: ['route', 'formats', 'section', 'matomo'] as unknown as undefined,
   setup(props) {
     const opened = ref<boolean>(false)
     const toggle = (newState: boolean) => {
@@ -28,22 +29,20 @@ export const PureDownloads = defineComponent<Props>({
               {props.formats.map(format => {
                 return (
                   <Download
+                    class="btn-alt small px-s py-s full-x border-b-s"
                     key={format}
                     format={format}
-                    class="btn-alt small px-s py-s full-x border-b-s"
                     section={props.section}
                     query={props.route.query}
-                    params={props.params}
-                    onClicked={toggle}
+                    matomo={props.matomo}
+                    onClicked={() => toggle(!opened.value)}
                   />
                 )
               })}
             </div>
           )
         }}
-      >
-        <div></div>
-      </Dropdown>
+      />
     )
   }
 })
