+++ packages/ui/src/components/_ui/chip.stories.ts
@@ -3,7 +3,7 @@ import { Meta, Story } from '@storybook/vue3'
 import { action } from '@storybook/addon-actions'
 
 const meta: Meta = {
-  title: 'UI/Chip',
+  title: 'Components/UI/Chip',
   component: Chip,
   argTypes: {}
 }
