+++ packages/ui/src/components/dashboard/pure-dgtm-stats.stories.ts
@@ -4,7 +4,7 @@ import { StatistiquesDGTM } from 'camino-common/src/statistiques'
 import { statistiquesDGTMFake } from './testData'
 
 const meta: Meta = {
-  title: 'Components/NoStoryshots/PureDGTMStats',
+  title: 'Components/Dashboard/DGTMStats',
   component: PureDGTMStats,
   argTypes: {
     getEntreprisesTitres: { name: 'function', required: true },
@@ -25,16 +25,22 @@ const Template: Story<Props> = (args: Props) => ({
   template: '<PureDGTMStats v-bind="args" />'
 })
 
-export const Ok = Template.bind({})
-Ok.args = {
-  getDgtmStats: () => Promise.resolve(statistiquesDGTMFake)
-}
+export const OkNoSnapshot = Template.bind(
+  {},
+  {
+    getDgtmStats: () => Promise.resolve(statistiquesDGTMFake)
+  }
+)
 
-export const Loading = Template.bind({})
-Loading.args = {
-  getDgtmStats: () => new Promise<StatistiquesDGTM>(resolve => {})
-}
-export const WithError = Template.bind({})
-WithError.args = {
-  getDgtmStats: () => Promise.reject(new Error('because reasons'))
-}
+export const Loading = Template.bind(
+  {},
+  {
+    getDgtmStats: () => new Promise<StatistiquesDGTM>(resolve => {})
+  }
+)
+export const WithError = Template.bind(
+  {},
+  {
+    getDgtmStats: () => Promise.reject(new Error('because reasons'))
+  }
+)
